requirements.txt
# Core dependencies
python-dotenv
pydantic==2.10.5

# Web framework (choose one or both if needed)

uvicorn==0.18.3 
flask==3.0.3


# LangChain ecosystem
langchain
langchain-community
langchain-openai
langchain-huggingface
langchain_experimental
-e .

# PDF parsing (for future: e.g., astrological texts or corpus)
pypdf
sentence-transformers==2.2.2

# Utility
requests==2.31.0


